requirements.txt
tensorflow
streamlit
pillow
matplotlib
fpdf
numpy
pandas
